Add files via upload
The purpose of this analysis is to develop a machine learning model that predicts the success of organizations applying for funding from Alphabet Soup, a nonprofit foundation. By doing so, it aims to improve the efficiency of funding decisions, enhance decision-making processes, and streamline the application process, ultimately enabling Alphabet Soup to allocate resources more effectively and make a greater impact through its philanthropic efforts.